refactor: split operator_commands.rs into module directory
Split the monolithic 2,364-line operator_commands.rs into a module
directory with 9 focused sub-modules:

  mod.rs        (51 lines)  - declarations and re-exports
  dispatch.rs   (600 lines) - dispatch_operator_probe, dispatch_legacy_gym_cli + tests
  probe.rs      (221 lines) - bootstrap, handoff, copilot-submit probes
  state_root.rs (312 lines) - state root resolution and parse_runtime_topology
  validation.rs (346 lines) - filesystem validation and read artifact helpers
  evidence.rs   (310 lines) - terminal evidence and objective helpers
  format.rs     (168 lines) - print_text, print_display, section formatters
  goals.rs      (218 lines) - GoalRegisterView and GoalRegisterSection
  recipes.rs    (193 lines) - terminal recipe loading and validation

All 364 operator_commands tests pass. No external API changes -
callers using crate::operator_commands::<item> continue to compile
without modification.
